Avoid the need of sleep interval, see details
Instead of each thread getting host from Queue and trying user-password
pairs on it with sleep in between of tries, we now iterate over Queue of
user-password pairs and in each thread try to apply it to each host.
Given the number of hosts - it will take time to try it at all hosts.